+++ src/auth/send_email.h
@@ -16,7 +16,7 @@ namespace celeritas
         [[nodiscard]] void_awaitable_type response();
 
     private:
-        [[nodiscard]] static std::string calculate_hmac_sha256(int app_id, const std::string& email, int64_t timestamp, const std::string& secret_key);
+        [[nodiscard]] static std::string calculate_hmac_sha256(int64_t app_id, const std::string& email, int64_t timestamp, const std::string& secret_key);
 
         void send_sdk_email(const email_code& email_code);
